ServiceUser now abstract

+++ TutoringService321-Backend/src/main/java/ca/mcgill/ecse321/tutoringservice321/service/TutoringService321Service.java
@@ -10,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import ca.mcgill.ecse321.tutoringservice321.TutoringService321Application;
 import ca.mcgill.ecse321.tutoringservice321.dao.*;
 import ca.mcgill.ecse321.tutoringservice321.model.*;
 
@@ -74,7 +75,7 @@ public class TutoringService321Service {
 	public List<Tutor> getAllTutors() {
 		return toList(tutorRepository.findAll());
 	}
-	
+
 	@Transactional
 	public Subject createSubject(String name) {
 		Subject subject = new Subject();
@@ -102,7 +103,7 @@ public class TutoringService321Service {
 	public List<Subject> getAllSubjects() {
 		return toList(subjectRepository.findAll());
 	}
-	
+
 	@Transactional
 	public Session createSession(String tutorEmail, Date date, Time endTime, Time startTime) {
 		//Find the tutor first
@@ -151,7 +152,7 @@ public class TutoringService321Service {
 	public List<Session> getAllSessions() {
 		return toList(sessionRepository.findAll());
 	}
-	
+
 	@Transactional
 	public Course createCourse(String description, String school, String courseCode) {
 		Course course = new Course();
@@ -192,12 +193,16 @@ public class TutoringService321Service {
 	public List<Course> getAllCourses() {
 		return toList(courseRepository.findAll());
 	}
-	
+
 	//====================================================================================
 	//AVAILABILITY METHODS
-	
+
 	@Transactional
 	public Availability addAvailability(String tutorEmail, Date date, Time startTime, Time endTime) {
+		if(TutoringService321Application.getLoggedUser() == null || !(TutoringService321Application.getLoggedUser() instanceof Tutor)) {
+			throw new IllegalArgumentException("A tutor must be logged in to perform this operation");
+		}
+		
 		//Find the tutor first
 		Tutor tutor = getTutor(tutorEmail);
 
@@ -214,7 +219,7 @@ public class TutoringService321Service {
 		if(tutor == null) {
 			throw new IllegalArgumentException("Tutor cannot be empty.");
 		}
-		
+
 		//Checking if this Availability already exists
 		List<Availability> availabilities = getAllTutorAvailabilities(tutorEmail);
 		for(Availability avail :  availabilities) {
@@ -222,7 +227,7 @@ public class TutoringService321Service {
 				throw new IllegalArgumentException("Availability already exists.");
 			}
 		}
-		
+
 		Availability availability = new Availability();
 		//Setting attributes
 		availability.setDate(date);
@@ -236,7 +241,8 @@ public class TutoringService321Service {
 		return availability;
 	}
 
-	public Availability getAvailability(String tutorEmail, Date date, Time startTime, Time endTime) {
+	@Transactional
+	public Availability getAvailability(String tutorEmail, Date date, Time startTime, Time endTime) {	
 		//Find the tutor first
 		Tutor tutor = getTutor(tutorEmail);
 
@@ -251,23 +257,28 @@ public class TutoringService321Service {
 		return null;
 	}
 
+	@Transactional
 	public void deleteAvailability(String tutorEmail, Date date, Time startTime, Time endTime) {
-		Availability availability = getAvailability(tutorEmail, date, startTime, endTime);
+		if(TutoringService321Application.getLoggedUser() == null || !(TutoringService321Application.getLoggedUser() instanceof Tutor)) {
+			throw new IllegalArgumentException("A tutor must be logged in to perform this operation");
+		}
 		
+		Availability availability = getAvailability(tutorEmail, date, startTime, endTime);
+
 		if(availability != null) {
 			availabilityRepository.delete(availability);
 		}
 	}
-	
+
 	@Transactional
 	public List<Availability> getAllTutorAvailabilities(String tutorEmail) {
 		//Find the tutor first
 		Tutor tutor = getTutor(tutorEmail);
-		
+
 		if(tutor == null) {
 			throw new IllegalArgumentException("There is no such Tutor.");
 		}
-		
+
 		List<Availability> list = toList(availabilityRepository.findAll());
 		List<Availability> tutorAvailabilities = new ArrayList<Availability>();
 		for(Availability availability : list) {
@@ -275,12 +286,17 @@ public class TutoringService321Service {
 				tutorAvailabilities.add(availability);
 			}
 		}
-		
+
 		return toList(tutor.getAvailability());
 	}
 
+	@Transactional
 	public Availability updateAvailability(String tutorEmail, Date oldDate, Time oldStartTime, Time oldEndTime,
 			Date newDate, Time newStartTime, Time newEndTime) {
+		if(TutoringService321Application.getLoggedUser() == null || !(TutoringService321Application.getLoggedUser() instanceof Tutor)) {
+			throw new IllegalArgumentException("A tutor must be logged in to perform this operation");
+		}
+		
 		//We first check that there is no session at that time
 		Tutor tutor = tutorRepository.findTutorByEmail(tutorEmail);
 		Set<Session> sessions = sessionRepository.findSessionByTutorAndDate(tutor, oldDate);
@@ -289,18 +305,45 @@ public class TutoringService321Service {
 				throw new IllegalArgumentException("Already an availability at that time and date.");
 			}
 		}
-		
+
 		//We first delete the old availability
 		deleteAvailability(tutorEmail, oldDate, oldStartTime, oldEndTime);
-		
+
 		//We then add the new one
 		Availability availability = addAvailability(tutorEmail, newDate, newStartTime, newEndTime);
-	
+
 		return availability;
+	}
+
+	//====================================================================================
+	//LOGIN-LOGOUT METHODS
+
+	@Transactional
+	public void loginAsTutor(String email, String password) {
+		//Input validation
+		if(email == null || email.trim().length() == 0) {
+			throw new IllegalArgumentException("Email cannot be empty.");
+		}
+		if(password == null || password.trim().length() == 0) {
+			throw new IllegalArgumentException("Password cannot be empty.");
+		}
+
+		List<Tutor> tutors = getAllTutors();
+
+		for(Tutor tutor : tutors) {
+			if(tutor.getEmail().equals(email) && tutor.getEmail().equals(password)) {
+				TutoringService321Application.setLoggedUser(tutor);
+			}
+		}
+	}
+
+	@Transactional
+	public void logout() {
+		TutoringService321Application.setLoggedUser(null);
 	}
 	
 	//====================================================================================
-	
+
 	//Helper method provided in EventRegistration
 	private <T> List<T> toList(Iterable<T> iterable){
 		List<T> resultList = new ArrayList<T>();
